BREAKING CHANGE: merge wave-cli, wave-metrics, api-server into wave-autoscale
