+++ .moon/archiver/archive-script.sh
@@ -15,14 +15,11 @@ mkdir -p archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
 # Copy the binaries
 cp -r target/$PLATFORM/release/wave-controller archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
 cp -r target/$PLATFORM/release/wave-api-server archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
-cp -r target/$PLATFORM/release/wave-metrics archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
-cp -r target/$PLATFORM/release/wave-cli archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
 cp -r core/web-app/.next/standalone archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM/wave-web-app
 cp -r core/web-app/.next/static archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM/wave-web-app/.next/static
 
 # Copy the configuration files
-cp -r tests/config/wave-config.yaml archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
-cp -r core/wave-metrics/tests/collectors/collectors.yaml archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
+cp -r core/wave-autoscale/tests/config/wave-config.yaml archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
 
 # Archive the binaries
 tar -C archive -czvf ./archive/wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM.tar.gz wave-autoscale-$VERSION/wave-autoscale-$VERSION-$PLATFORM
