+++ .moon/archiver/moon.yml
@@ -17,10 +17,8 @@ tasks:
       - "**/*"
       - "./archive-script.sh"
     deps:
-      - "wave-metrics:build-aarch64-apple-darwin"
       - "api-server:build-aarch64-apple-darwin"
       - "controller:build-aarch64-apple-darwin"
-      - "cli:build-aarch64-apple-darwin"
       # - "web-app:build"
     env:
       PLATFORM: "aarch64-apple-darwin"
@@ -30,10 +28,8 @@ tasks:
     inputs:
       - "**/*"
     deps:
-      - "wave-metrics:build-x86_64-unknown-linux-gnu"
       - "api-server:build-x86_64-unknown-linux-gnu"
       - "controller:build-x86_64-unknown-linux-gnu"
-      - "cli:build-x86_64-unknown-linux-gnu"
       # - "web-app:build"
     env:
       PLATFORM: "x86_64-unknown-linux-gnu"
