+++ core/api-server/src/app.rs
@@ -2,7 +2,7 @@ use crate::{app_state::get_app_state, controller};
 use actix_cors::Cors;
 use actix_web::{App, HttpServer};
 use data_layer::data_layer::DataLayer;
-use log::info;
+use log::{debug, info};
 use std::{
     sync::Arc,
     time::{SystemTime, UNIX_EPOCH},
@@ -26,9 +26,11 @@ pub async fn run_api_server(
     let host = wave_config.host;
     let port = wave_config.port;
 
+    debug!("host: {}", host);
     // Run HTTP Server
     let app_state = get_app_state(shared_data_layer);
 
+    debug!("app_state");
     let http_server = HttpServer::new(move || {
         let cors = Cors::permissive().max_age(3600);
 
@@ -43,16 +45,19 @@ pub async fn run_api_server(
             .configure(controller::init_autoscaling_history_controller)
             .configure(controller::init_metrics_receiver_controller)
     })
+    .workers(1)
     .bind((host.clone(), port));
 
-    // Server structure implements Future.
-    let server = match http_server {
-        Ok(server) => server.run(),
-        Err(error) => {
-            panic!("There was a problem opening the file: {:?}", error)
-        }
-    };
+    if http_server.is_err() {
+        panic!(
+            "There was a problem opening the file: {:?}",
+            http_server.err().unwrap()
+        )
+    }
+    let http_server = http_server.unwrap();
+    let _ = http_server.run().await;
 
     info!("It's up! {}:{}", host, port);
-    server.await
+    // server.await
+    Ok(())
 }
