+++ core/wave-autoscale/moon.yml
@@ -79,8 +79,8 @@ tasks:
       - "--"
       - "--config"
       - "./tests/config/wave-config.yaml"
-      - "--definition"
-      - "./core/wave-autoscale/tests/yaml/plan_cloudwatch_ecs_ec2.yaml"
+      # - "--definition"
+      # - "./core/wave-autoscale/tests/yaml/plan_cloudwatch_ecs_ec2.yaml"
     env:
       RUST_LOG: wave_autoscale=debug
   run-help:
