+++ core/wave-autoscale/src/lib.rs
@@ -5,3 +5,4 @@ pub mod metric_updater;
 pub mod scaling_component;
 pub mod scaling_planner;
 pub mod util;
+pub mod web_app_runner;
