+++ core/wave-autoscale/tests/config/wave-config.yaml
@@ -1,3 +1,12 @@
+db_url: sqlite://./wave.db
+host: localhost
+port: 3024
+watch_definition_duration: 5000
+autoscaling_history_retention: '1d'
+web_ui: false
+web_ui_host: 0.0.0.0
+web_ui_port: 3025
+
 vector:
   macos_x86_64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-apple-darwin.tar.gz
   macos_aarch64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-apple-darwin.tar.gz
