+++ core/wave-autoscale/tests/metrics-examples/telegraf-cloudwatch-metrics.yaml
