+++ core/wave-autoscale/tests/metrics-examples/telegraf-google-cloud-monitoring-metrics.yaml
