+++ core/wave-autoscale/tests/metrics-examples/telegraf-mem-metrics.yaml
