+++ core/wave-autoscale/tests/metrics-examples/vector-host-metrics.yaml
