+++ core/wave-autoscale/tests/metrics-examples/vector-prometheus-metrics.yaml
