--- core/wave-cli/moon.yml
@@ -1,115 +1,0 @@
-language: "rust"
-type: "application"
-owners:
-  defaultOwner: "@wave-autoscale-team"
-
-tasks:
-  test:
-    command: "cargo nextest run -p wave-cli"
-    env:
-      RUST_LOG: wave_cli=debug
-  copy-wave-controller:
-    command: "cp ../../target/aarch64-apple-darwin/release/wave-controller ./wave-controller"
-    platform: "system"
-    deps:
-      - "controller:build-aarch64-apple-darwin"
-    outputs:
-      - "./wave-controller"
-  copy-wave-metrics:
-    command: "cp ../../target/aarch64-apple-darwin/release/wave-metrics ./wave-metrics"
-    platform: "system"
-    deps:
-      - "wave-metrics:build-aarch64-apple-darwin"
-    outputs:
-      - "./wave-metrics"
-  copy-wave-api-server:
-    command: "cp ../../target/aarch64-apple-darwin/release/wave-api-server ./wave-api-server"
-    platform: "system"
-    deps:
-      - "api-server:build-aarch64-apple-darwin"
-    outputs:
-      - "./wave-api-server"
-  run:
-    command:
-      - "cargo"
-      - "run"
-      - "-p"
-      - "wave-cli"
-      - "--"
-      - "--run-metrics"
-      - "--run-api-server"
-      - "--config"
-      - "../../tests/config/wave-config.yaml"
-      - "--definition"
-      - "../wave-metrics/tests/metrics-examples/vector-host-metrics.yaml"
-      - "--collectors-info"
-      - "../wave-metrics/tests/collectors/collectors.yaml"
-    deps:
-      # TODO: Build web-app
-      - "copy-wave-controller"
-      - "copy-wave-metrics"
-      - "copy-wave-api-server"
-    outputs:
-      - "wave-controller"
-      - "wave-api-server"
-      - "wave-metrics"
-      - "wave-web-app/**/*"
-      - "tests/db/wave.db"
-    env:
-      RUST_LOG: debug
-    options:
-      runFromWorkspaceRoot: false
-  run-help:
-    command:
-      - "cargo"
-      - "run"
-      - "-p"
-      - "wave-cli"
-      - "--"
-      - "--help"
-    env:
-      RUST_LOG: wave_cli=debug
-    options:
-      runFromWorkspaceRoot: false
-
-  # Build for Production
-  # Linux
-  build-aarch64-unknown-linux-gnu:
-    command: "cargo build --release --target aarch64-unknown-linux-gnu"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-aarch64-unknown-linux-musl:
-    command: "cargo build --release --target aarch64-unknown-linux-musl"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-x86_64-unknown-linux-gnu:
-    command: "cargo build --release --target x86_64-unknown-linux-gnu"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-x86_64-unknown-linux-musl:
-    command: "cargo build --release --target x86_64-unknown-linux-musl"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  # Mac
-  build-x86_64-apple-darwin:
-    command: "cargo build --release --target x86_64-apple-darwin"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-aarch64-apple-darwin:
-    command: "cargo build --release --target aarch64-apple-darwin"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-    outputs:
-      - "/target/aarch64-apple-darwin/release/wave-cli"
