--- core/wave-cli/rust-toolchain.toml
@@ -1,2 +1,0 @@
-[toolchain]
-channel = "1.70.0"
