--- core/wave-cli/tests/integration_test.rs
@@ -1,6 +1,0 @@
-mod integration_test {
-    #[test]
-    fn test_() {
-        // TODO: Implement integration test
-    }
-}
