--- core/wave-cli/tests/yaml/definition.yaml
@@ -1,28 +1,0 @@
-# References - Telegraf Cloudwatch Input Plugin
-# https://github.com/influxdata/telegraf/blob/master/plugins/inputs/cloudwatch/README.md
----
-kind: Metric
-id: cloudwatch_dynamodb_metric
-collector: telegraf
-metric_kind: cloudwatch
-metadata:
-  region: ap-northeast-3
-  # AWS Credentials
-  profile: "default"
-
-  # Requested CloudWatch aggregation Period (required)
-  # Must be a multiple of 60s.
-  period: "2m"
-  # Collection Delay (required)
-  # Must account for metrics availability via CloudWatch API
-  delay: "2m"
-  # Recommended: use metric 'interval' that is a multiple of 'period' to avoid
-  # gaps or overlap in pulled data
-  interval: "1m"
-  namespaces: ["AWS/DynamoDB"]
-  metrics:
-    - names: ["ConsumedWriteCapacityUnits"]
-      statistic_include: ["sample_count"]
-      dimensions:
-        - name: TableName
-          value: "test-dynamodb-table"
