--- core/wave-metrics/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "wave-metrics"
-version = "0.1.0"
-edition = "2021"
-
-[[bin]]
-name = "wave-metrics"
-path = "src/main.rs"
-
-[profile.release]
-# Tell `rustc` to optimize for small code size. https://nnethercote.github.io/perf-book/build-configuration.html#link-time-optimization
-lto = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# Internal dependencies
-data-layer = { path = "../data-layer" }
-utils = { path = "../utils" }
-# 3rd party dependencies
-clap = { version = "4.3.0", features = ["derive"] }
-env_logger = { version = "0.10.0" }
-log = { version = "0.4.18" }
-anyhow = "1.0.70"
-serde = { version = "1.0.157", features = ["derive"] }
-serde_json = "1.0.94"
-serde_yaml = "0.9.19"
-reqwest = { version = "0.11.15", features = ["json", "stream"] }
-futures-util = "0.3.14"
-tokio = { version = "1.27.0", features = ["full"] }
-flate2 = { version = "1.0.26" }
-tar = { version = "0.4.38" }
-walkdir = { version = "2.3.3" }
-toml = { version = "0.7.4" }
