--- core/wave-metrics/moon.yml
@@ -1,90 +1,0 @@
-language: "rust"
-type: "application"
-owners:
-  defaultOwner: "@wave-autoscale-team"
-
-dependsOn:
-  - "data-layer"
-  - "utils"
-
-tasks:
-  test:
-    command: "cargo nextest run -p wave-metrics --test-threads 1"
-    deps:
-      - "lint"
-    env:
-      RUST_LOG: wave_metrics=debug
-  test-metric-collector-manager:
-    command: "cargo nextest run -p wave-metrics metric_collector_manager:: --test-threads 1"
-    deps:
-      - "lint"
-    env:
-      RUST_LOG: wave_metrics=debug
-  run:
-    command:
-      - "cargo"
-      - "run"
-      - "-p"
-      - "wave-metrics"
-      - "--"
-      - "--config"
-      - "./tests/config/wave-config.yaml"
-      - "--definition"
-      - "./core/wave-metrics/tests/metrics-examples/vector-host-metrics.yaml"
-      - "--collectors-info"
-      - "./core/wave-metrics/tests/collectors/collectors.yaml"
-    env:
-      RUST_LOG: wave_metrics=debug
-    inputs:
-      - "@globs(tests)"
-  run-help:
-    command:
-      - "cargo"
-      - "run"
-      - "-p"
-      - "wave-metrics"
-      - "--"
-      - "--help"
-    env:
-      RUST_LOG: wave_metrics=debug
-  # Build for Production
-  # Linux
-  build-aarch64-unknown-linux-gnu:
-    command: "cargo build --release --target aarch64-unknown-linux-gnu"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-aarch64-unknown-linux-musl:
-    command: "cargo build --release --target aarch64-unknown-linux-musl"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-x86_64-unknown-linux-gnu:
-    command: "cargo build --release --target x86_64-unknown-linux-gnu"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-x86_64-unknown-linux-musl:
-    command: "cargo build --release --target x86_64-unknown-linux-musl"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  # Mac
-  build-x86_64-apple-darwin:
-    command: "cargo build --release --target x86_64-apple-darwin"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-  build-aarch64-apple-darwin:
-    command: "cargo build --release --target aarch64-apple-darwin"
-    deps:
-      - "test"
-    inputs:
-      - "@globs(sources)"
-    outputs:
-      - "/target/aarch64-apple-darwin/release/wave-metrics"
