--- core/wave-metrics/tests/collectors/collectors.yaml
@@ -1,12 +1,0 @@
-vector:
-  macos_x86_64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-apple-darwin.tar.gz
-  macos_aarch64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-apple-darwin.tar.gz
-  linux_x86_64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-unknown-linux-gnu.tar.gz
-  linux_aarch64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-aarch64-unknown-linux-gnu.tar.gz
-  windows_x86_64: https://github.com/vectordotdev/vector/releases/download/v0.30.0/vector-0.30.0-x86_64-pc-windows-msvc.zip
-telegraf:
-  macos_x86_64: https://dl.influxdata.com/telegraf/releases/telegraf-1.27.1_darwin_amd64.tar.gz
-  macos_aarch64: https://dl.influxdata.com/telegraf/releases/telegraf-1.27.1_darwin_arm64.tar.gz
-  linux_x86_64: https://dl.influxdata.com/telegraf/releases/telegraf-1.27.1_linux_amd64.tar.gz
-  linux_aarch64: https://dl.influxdata.com/telegraf/releases/telegraf-1.27.1_linux_arm64.tar.gz
-  windows_x86_64: https://dl.influxdata.com/telegraf/releases/telegraf-1.27.1_windows_amd64.zip
