--- tests/config/wave-config.yaml
@@ -1,14 +1,0 @@
-common:
-  db_url: sqlite://./wave.db
-
-wave_metrics:
-  output:
-    url: http://localhost:3024/api/metrics-receiver
-
-wave_api_server:
-  port: 3024
-  host: localhost
-
-wave_web_app:
-  port: 3004
-  host: localhost
